requirements.txt
# Library dependencies for the python code.  You need to install these with
# `pip install -r requirements.txt` before you can run this.

selenium
lxml
requests
python-dotenv
pyautogui